+++ tests/test.py
@@ -383,9 +383,7 @@ def get_image_paths(folder_path):
                    if os.path.splitext(file)[1].lower() in image_extensions]
     return image_paths
 if __name__ == "__main__":
-    os.environ["PI_SERVER_URL"] = "https://api.prodv2.cryptonets.ai/node"
-    os.environ["PI_API_KEY"] = "00000000000000001962"
-   
+
     (face_factor, image_path_1,) = setup_test(
         "test.jpeg",
     )
@@ -409,15 +407,15 @@ if __name__ == "__main__":
         
 
     # test_age_estimate_with_no_cache()
-    compare_start_time = default_timer()
-    compare_handle = face_factor.compare(image_path_1="/home/azam/projects/openinfer/python sdk/cryptonets-python-sdk/docs/source/Usage/images/tom_hanks.png",image_path_2="/home/azam/projects/openinfer/python sdk/cryptonets-python-sdk/docs/source/Usage/images/tom_hanks_2.jpg")
-    print("Duration:", default_timer() - compare_start_time, "\n")
-    print("Status:{}\nResult:{}\nMessage:{}\nMin:{}\nMean:{}\nMax:{}\n1VR:{}\n2VR:{}\n".format(
-        compare_handle.status,
-        compare_handle.result, compare_handle.message, compare_handle.distance_min,
-        compare_handle.distance_mean,
-        compare_handle.distance_max, compare_handle.first_validation_result,
-        compare_handle.second_validation_result))
+    # compare_start_time = default_timer()
+    # compare_handle = face_factor.compare(image_path_1=img1,image_path_2=img2)
+    # print("Duration:", default_timer() - compare_start_time, "\n")
+    # print("Status:{}\nResult:{}\nMessage:{}\nMin:{}\nMean:{}\nMax:{}\n1VR:{}\n2VR:{}\n".format(
+    #     compare_handle.status,
+    #     compare_handle.result, compare_handle.message, compare_handle.distance_min,
+    #     compare_handle.distance_mean,
+    #     compare_handle.distance_max, compare_handle.first_validation_result,
+    #     compare_handle.second_validation_result))
     # test_compare_with_no_cache()
     # test_get_iso_image_with_cache()
     # test_get_iso_image_with_no_cache()